+++ agate/aggregations/max_length.py
@@ -11,6 +11,10 @@ class MaxLength(Aggregation):
     """
     Find the length of the longest string in a column.
 
+    Note: On Python 2.7 this function may miscalcuate the length of unicode
+    strings that contain "wide characters". For details see this StackOverflow
+    answer: http://stackoverflow.com/a/35462951
+
     :param column_name:
         The name of a column containing :class:`.Text` data.
     """